scripts/requirements.txt
anthropic>=0.40.0
requests==2.31.0
psycopg2-binary==2.9.9
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-secret-manager==2.26.0
